double g2p/p2g (developing)

+++ functional/src/voxel.cuh
@@ -3,31 +3,31 @@
 
 // CUDA function declarations
 void avg_voxelize(int b, int c, int n, int r, int r2, int r3, const int *coords,
-                  const float *feat, int *ind, int *cnt, float *out);
+                  const double *feat, int *ind, int *cnt, double *out);
 void avg_voxelize_grad(int b, int c, int n, int s, const int *idx,
-                       const int *cnt, const float *grad_y, float *grad_x);
+                       const int *cnt, const double *grad_y, double *grad_x);
 
 
 void mpm_point2voxel(
-    float *__restrict__ xyz, //n x 3
-    float *__restrict__ feature, //n x c 
+    double *__restrict__ xyz, //n x 3
+    double *__restrict__ feature, //n x c 
     int gx, int gy, int gz, // dx, dy, dz
-    float dx,
-    float *__restrict__ voxel,
+    double dx,
+    double *__restrict__ voxel,
     int *__restrict__ batch_index,
     int d,
     int c,
     int dim);
 
 void mpm_point2voxel_grad(
-    float *__restrict__ xyz, //n x 3
-    float *__restrict__ feature, //n x c 
+    double *__restrict__ xyz, //n x 3
+    double *__restrict__ feature, //n x c 
     int gx, int gy, int gz,
-    float dx,
-    float *__restrict__ voxel,
-    float *__restrict__ voxel_grad,
-    float *__restrict__ x_grad,
-    float *__restrict__ feature_grad,
+    double dx,
+    double *__restrict__ voxel,
+    double *__restrict__ voxel_grad,
+    double *__restrict__ x_grad,
+    double *__restrict__ feature_grad,
     int *__restrict__ batch_index,
     int d,
     int c,
